Issue #6: Refactor task structure.
