+++ grumphp.yml.dist
@@ -12,6 +12,6 @@ parameters:
         - php
     git_commit_message:
       matchers:
-        - '/^(([A-Z]+\-\d+)|(#\d+)): [A-Z].+\./'
+        - '/^([A-Z]+\-\d+|Issue #\d+): [A-Z].+\./'
       case_insensitive: false
       multiline: false
