+++ tests/TaskTest.php
@@ -3,7 +3,7 @@
 namespace NuvoleWeb\Robo\Tests;
 
 use League\Container\ContainerAwareInterface;
-use NuvoleWeb\Robo\Task\Config\AppendConfiguration;
+use NuvoleWeb\Robo\Task\Config\Php\AppendConfiguration;
 use NuvoleWeb\Robo\Task\Config\loadTasks;
 use PHPUnit\Framework\TestCase;
 use Robo\Config\Config;
